Move `ModelMonitoringSchedulesFile` to a module
Including the tests
